Make CRUD methods transaction compatible
All steps of a transaction must be performed on the same connection.
To make this possible, connections used by CRUD methods do not
automatically get released but beginTransaction returns its conn
instead so it can be passed to the other methods.

If commit or rollback are not called, the connection will be lost
to the pool though!